Team Ctrl_FY_1
Our team (Team Ctrl) has created an interface named "ClubGenie" as a part of the "Open Innovation" project for Buffer 4.0. The interface is built using Java and Eclipse IDE. The main objective of the interface is to act as a guide for the various student clubs present in MKSSS's Cummins College of Engineering.

ClubGenie offers three main functionalities. Firstly, it displays a list of different clubs present in the college and provides detailed information about each one of them. Secondly, it offers personalized club suggestions to the user by asking them to fill out a questionnaire. Lastly, it displays information about past events conducted by the clubs and upcoming events planned ahead.

The project was created using arraylist, string, switch case, and conditional statements. It was a great learning experience for our team, especially as we are from the first year of college. We had the opportunity to explore DSA for the first time and work on a project that involved programming concepts.

Overall, ClubGenie is a useful and efficient interface that can provide information to students about different clubs and events in the college. The Google Drive link containing all the required documents for the project is attached below.
https://drive.google.com/drive/folders/1ODYUlQ4-i7AWFdkS66YiZLJEBI28K-xg?usp=share_link